lookupKey working
