+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OzoneManager.java
@@ -431,7 +431,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
 
   private final boolean isSecurityEnabled;
 
-  private SnapshotManager snapshotManager;
+  private OmMReader omMReader;
 
   
   @SuppressWarnings("methodlength")
@@ -733,23 +733,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
       instantiatePrepareStateOnStartup();
     }
 
-    if (isAclEnabled) {
-      accessAuthorizer = getACLAuthorizerInstance(configuration);
-      if (accessAuthorizer instanceof OzoneNativeAuthorizer) {
-        OzoneNativeAuthorizer authorizer =
-            (OzoneNativeAuthorizer) accessAuthorizer;
-        isNativeAuthorizerEnabled = true;
-        authorizer.setVolumeManager(volumeManager);
-        authorizer.setBucketManager(bucketManager);
-        authorizer.setKeyManager(keyManager);
-        authorizer.setPrefixManager(prefixManager);
-        authorizer.setOzoneAdmins(omAdminUsernames);
-        authorizer.setAllowListAllVolumes(allowListAllVolumes);
-      }
-    } else {
-      accessAuthorizer = null;
-    }
-    snapshotManager = new SnapshotManager(keyManager, prefixManager, volumeManager, bucketManager, metadataManager, configuration, omRpcAddress);
+    omMReader = new OmMReader(keyManager, prefixManager,  metadataManager, this);
   }
 
   /**
@@ -838,22 +822,6 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
     KeyProviderCryptoExtension cryptoProvider = KeyProviderCryptoExtension
         .createKeyProviderCryptoExtension(keyProvider);
     return cryptoProvider;
-  }
-
-  /**
-   * Returns an instance of {@link IAccessAuthorizer}.
-   * Looks up the configuration to see if there is custom class specified.
-   * Constructs the instance by passing the configuration directly to the
-   * constructor to achieve thread safety using final fields.
-   *
-   * @param conf
-   * @return IAccessAuthorizer
-   */
-  private IAccessAuthorizer getACLAuthorizerInstance(OzoneConfiguration conf) {
-    Class<? extends IAccessAuthorizer> clazz = conf.getClass(
-        OZONE_ACL_AUTHORIZER_CLASS, OzoneAccessAuthorizer.class,
-        IAccessAuthorizer.class);
-    return ReflectionUtils.newInstance(clazz, conf);
   }
 
   @Override
@@ -1412,8 +1380,16 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
     return prefixManager;
   }
 
+  public VolumeManager getVolumeManager() {
+    return volumeManager;
+  }
+
+  public BucketManager getBucketManager() {
+    return bucketManager;
+  }
+
   public IAccessAuthorizer getAccessAuthorizer() {
-    return accessAuthorizer;
+    return omMReader.getAccessAuthorizer();
   }
 
   /**
@@ -2353,40 +2329,6 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
     }
   }
 
-  /**
-   * Checks if current caller has acl permissions.
-   *
-   * @param resType - Type of ozone resource. Ex volume, bucket.
-   * @param store   - Store type. i.e Ozone, S3.
-   * @param acl     - type of access to be checked.
-   * @param vol     - name of volume
-   * @param bucket  - bucket name
-   * @param key     - key
-   * @throws OMException ResultCodes.PERMISSION_DENIED if permission denied.
-   */
-  private void checkAcls(ResourceType resType, StoreType store,
-      ACLType acl, String vol, String bucket, String key)
-      throws IOException {
-    UserGroupInformation user;
-    if (getS3Auth() != null) {
-      String principal =
-          OzoneAclUtils.accessIdToUserPrincipal(getS3Auth().getAccessId());
-      user = UserGroupInformation.createRemoteUser(principal);
-    } else {
-      user = ProtobufRpcEngine.Server.getRemoteUser();
-    }
-
-    InetAddress remoteIp = ProtobufRpcEngine.Server.getRemoteIp();
-    String volumeOwner = getVolumeOwner(vol, acl, resType);
-    String bucketOwner = getBucketOwner(vol, bucket, acl, resType);
-
-    OzoneAclUtils.checkAllAcls(this, resType, store, acl,
-        vol, bucket, key, volumeOwner, bucketOwner,
-        user != null ? user : getRemoteUser(),
-        remoteIp != null ? remoteIp : omRpcAddress.getAddress(),
-        remoteIp != null ? remoteIp.getHostName() : omRpcAddress.getHostName());
-  }
-
   public boolean isOwner(UserGroupInformation callerUgi, String ownerName) {
     if (ownerName == null) {
       return false;
@@ -2410,91 +2352,10 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
           UserGroupInformation.createRemoteUser(userName),
           ProtobufRpcEngine.Server.getRemoteIp(),
           ProtobufRpcEngine.Server.getRemoteIp().getHostName(),
-          false, getVolumeOwner(vol, acl, resType));
+          false, omMReader.getVolumeOwner(vol, acl, resType));
     } catch (OMException ex) {
       // Should not trigger exception here at all
       return false;
-    }
-  }
-
-  public String getVolumeOwner(String vol, ACLType type, ResourceType resType)
-      throws OMException {
-    String volOwnerName = null;
-    if (!vol.equals(OzoneConsts.OZONE_ROOT) &&
-        !(type == ACLType.CREATE && resType == ResourceType.VOLUME)) {
-      volOwnerName = getVolumeOwner(vol);
-    }
-    return volOwnerName;
-  }
-
-  private String getVolumeOwner(String volume) throws OMException {
-    Boolean lockAcquired = metadataManager.getLock().acquireReadLock(
-        VOLUME_LOCK, volume);
-    String dbVolumeKey = metadataManager.getVolumeKey(volume);
-    OmVolumeArgs volumeArgs = null;
-    try {
-      volumeArgs = metadataManager.getVolumeTable().get(dbVolumeKey);
-    } catch (IOException ioe) {
-      if (ioe instanceof OMException) {
-        throw (OMException)ioe;
-      } else {
-        throw new OMException("getVolumeOwner for Volume " + volume + " failed",
-            ResultCodes.INTERNAL_ERROR);
-      }
-    } finally {
-      if (lockAcquired) {
-        metadataManager.getLock().releaseReadLock(VOLUME_LOCK, volume);
-      }
-    }
-    if (volumeArgs != null) {
-      return volumeArgs.getOwnerName();
-    } else {
-      throw new OMException("Volume " + volume + " is not found",
-          OMException.ResultCodes.VOLUME_NOT_FOUND);
-    }
-  }
-
-  /**
-   * Return the owner of a given bucket.
-   *
-   * @return String
-   */
-  public String getBucketOwner(String volume, String bucket, ACLType type,
-       ResourceType resType) throws OMException {
-    String bucketOwner = null;
-    if ((resType != ResourceType.VOLUME) &&
-        !(type == ACLType.CREATE && resType == ResourceType.BUCKET)) {
-      bucketOwner = getBucketOwner(volume, bucket);
-    }
-    return bucketOwner;
-  }
-
-  private String getBucketOwner(String volume, String bucket)
-      throws OMException {
-
-    Boolean lockAcquired = metadataManager.getLock().acquireReadLock(
-            BUCKET_LOCK, volume, bucket);
-    String dbBucketKey = metadataManager.getBucketKey(volume, bucket);
-    OmBucketInfo bucketInfo = null;
-    try {
-      bucketInfo = metadataManager.getBucketTable().get(dbBucketKey);
-    } catch (IOException ioe) {
-      if (ioe instanceof OMException) {
-        throw (OMException)ioe;
-      } else {
-        throw new OMException("getBucketOwner for Bucket " + volume + "/" +
-            bucket  + " failed: " + ioe.getMessage(),
-            ResultCodes.INTERNAL_ERROR);
-      }
-    } finally {
-      if (lockAcquired) {
-        metadataManager.getLock().releaseReadLock(BUCKET_LOCK, volume, bucket);
-      }
-    }
-    if (bucketInfo != null) {
-      return bucketInfo.getOwner();
-    } else {
-      throw new OMException("Bucket not found", ResultCodes.BUCKET_NOT_FOUND);
     }
   }
 
@@ -2526,42 +2387,8 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
         .setOwnerName(owner)
         .build();
 
-    return checkAcls(obj, context, throwIfPermissionDenied);
+    return omMReader.checkAcls(obj, context, throwIfPermissionDenied);
   }
-
-  /**
-   * CheckAcls for the ozone object.
-   *
-   * @return true if permission granted, false if permission denied.
-   * @throws OMException ResultCodes.PERMISSION_DENIED if permission denied
-   *                     and throwOnPermissionDenied set to true.
-   */
-  public boolean checkAcls(OzoneObj obj, RequestContext context,
-                           boolean throwIfPermissionDenied)
-      throws OMException {
-
-    if (!accessAuthorizer.checkAccess(obj, context)) {
-      if (throwIfPermissionDenied) {
-        String volumeName = obj.getVolumeName() != null ?
-                "Volume:" + obj.getVolumeName() + " " : "";
-        String bucketName = obj.getBucketName() != null ?
-                "Bucket:" + obj.getBucketName() + " " : "";
-        String keyName = obj.getKeyName() != null ?
-                "Key:" + obj.getKeyName() : "";
-        LOG.warn("User {} doesn't have {} permission to access {} {}{}{}",
-            context.getClientUgi().getUserName(), context.getAclRights(),
-            obj.getResourceType(), volumeName, bucketName, keyName);
-        throw new OMException("User " + context.getClientUgi().getUserName() +
-            " doesn't have " + context.getAclRights() +
-            " permission to access " + obj.getResourceType() + " " +
-            volumeName  + bucketName + keyName, ResultCodes.PERMISSION_DENIED);
-      }
-      return false;
-    } else {
-      return true;
-    }
-  }
-
 
 
   /**
@@ -2571,6 +2398,14 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
    */
   public boolean getAclsEnabled() {
     return isAclEnabled;
+  }
+
+  public boolean getAllowListAllVolumes() {
+    return allowListAllVolumes;
+  }
+
+  public OmMReader getOmMReader() {
+    return omMReader;
   }
 
   /**
@@ -2592,7 +2427,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
   @Override
   public OmVolumeArgs getVolumeInfo(String volume) throws IOException {
     if (isAclEnabled) {
-      checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.READ, volume,
+      omMReader.checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.READ, volume,
           null, null);
     }
 
@@ -2701,7 +2536,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
       if (!allowListAllVolumes) {
         // Only admin can list all volumes when disallowed in config
         if (isAclEnabled) {
-          checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.LIST,
+          omMReader.checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.LIST,
               OzoneConsts.OZONE_ROOT, null, null);
         }
       }
@@ -2728,7 +2563,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
       String startKey, String prefix, int maxNumOfBuckets)
       throws IOException {
     if (isAclEnabled) {
-      checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.LIST, volumeName,
+      omMReader.checkAcls(ResourceType.VOLUME, StoreType.OZONE, ACLType.LIST, volumeName,
           null, null);
     }
     boolean auditSuccess = true;
@@ -2767,7 +2602,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
   public OmBucketInfo getBucketInfo(String volume, String bucket)
       throws IOException {
     if (isAclEnabled) {
-      checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.READ, volume,
+      omMReader.checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.READ, volume,
           bucket, null);
     }
     boolean auditSuccess = true;
@@ -2802,17 +2637,17 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
   @Override
   public OmKeyInfo lookupKey(OmKeyArgs args) throws IOException {
     LOG.info("gbjLookupKey");
-    return snapshotManager.lookupKey(args);
+    return omMReader.lookupKey(args);
   }
 
   @Override
   public List<OmKeyInfo> listKeys(String volumeName, String bucketName,
       String startKey, String keyPrefix, int maxKeys) throws IOException {
 
-    ResolvedBucket bucket = resolveBucketLink(Pair.of(volumeName, bucketName));
+    ResolvedBucket bucket = omMReader.resolveBucketLink(Pair.of(volumeName, bucketName));
 
     if (isAclEnabled) {
-      checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.LIST,
+      omMReader.checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.LIST,
           bucket.realVolume(), bucket.realBucket(), keyPrefix);
     }
 
@@ -2848,7 +2683,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
     // bucket links not supported
 
     if (isAclEnabled) {
-      checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.LIST,
+      omMReader.checkAcls(ResourceType.BUCKET, StoreType.OZONE, ACLType.LIST,
           volumeName, bucketName, keyPrefix);
     }
 
@@ -3319,7 +3154,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
       final String bucketName, String keyName, String uploadID,
       int partNumberMarker, int maxParts)  throws IOException {
 
-    ResolvedBucket bucket = resolveBucketLink(Pair.of(volumeName, bucketName));
+    ResolvedBucket bucket = omMReader.resolveBucketLink(Pair.of(volumeName, bucketName));
 
     Map<String, String> auditMap = bucket.audit();
     auditMap.put(OzoneConsts.KEY, keyName);
@@ -3348,7 +3183,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
   public OmMultipartUploadList listMultipartUploads(String volumeName,
       String bucketName, String prefix) throws IOException {
 
-    ResolvedBucket bucket = resolveBucketLink(Pair.of(volumeName, bucketName));
+    ResolvedBucket bucket = omMReader.resolveBucketLink(Pair.of(volumeName, bucketName));
 
     Map<String, String> auditMap = bucket.audit();
     auditMap.put(OzoneConsts.PREFIX, prefix);
@@ -3373,7 +3208,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
 
   @Override
   public OzoneFileStatus getFileStatus(OmKeyArgs args) throws IOException {
-    ResolvedBucket bucket = resolveBucketLink(args);
+    ResolvedBucket bucket = omMReader.resolveBucketLink(args);
 
     boolean auditSuccess = true;
     Map<String, String> auditMap = bucket.audit(args.toAuditMap());
@@ -3406,11 +3241,11 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
 
   @Override
   public OmKeyInfo lookupFile(OmKeyArgs args) throws IOException {
-    ResolvedBucket bucket = resolveBucketLink(args);
+    ResolvedBucket bucket = omMReader.resolveBucketLink(args);
 
     if (isAclEnabled) {
       LOG.info("gbjAcl2");
-      snapshotManager.checkAcls(ResourceType.KEY, StoreType.OZONE, ACLType.READ,
+      omMReader.checkAcls(ResourceType.KEY, StoreType.OZONE, ACLType.READ,
           bucket.realVolume(), bucket.realBucket(), args.getKeyName());
     }
 
@@ -3447,10 +3282,10 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
       String startKey, long numEntries, boolean allowPartialPrefixes)
       throws IOException {
 
-    ResolvedBucket bucket = resolveBucketLink(args);
+    ResolvedBucket bucket = omMReader.resolveBucketLink(args);
 
     if (isAclEnabled) {
-      checkAcls(getResourceType(args), StoreType.OZONE, ACLType.READ,
+      omMReader.checkAcls(getResourceType(args), StoreType.OZONE, ACLType.READ,
           bucket.realVolume(), bucket.realBucket(), args.getKeyName());
     }
 
@@ -3489,7 +3324,7 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
 
     try {
       if (isAclEnabled) {
-        checkAcls(obj.getResourceType(), obj.getStoreType(), ACLType.READ_ACL,
+        omMReader.checkAcls(obj.getResourceType(), obj.getStoreType(), ACLType.READ_ACL,
             obj.getVolumeName(), obj.getBucketName(), obj.getKeyName());
       }
       metrics.incNumGetAcl();
@@ -4015,15 +3850,6 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
     }
   }
 
-  /**
-   * Returns true if OzoneNativeAuthorizer is enabled and false if otherwise.
-   *
-   * @return if native authorizer is enabled.
-   */
-  public boolean isNativeAuthorizerEnabled() {
-    return isNativeAuthorizerEnabled;
-  }
-
   @VisibleForTesting
   public boolean isRunning() {
     return omState == State.RUNNING;
@@ -4042,94 +3868,19 @@ public final class OzoneManager extends ServiceRuntimeInfoImpl
         Pair.of(args.getVolumeName(), args.getBucketName()), omClientRequest);
   }
 
-  public ResolvedBucket resolveBucketLink(OmKeyArgs args)
-      throws IOException {
-    return resolveBucketLink(
-        Pair.of(args.getVolumeName(), args.getBucketName()));
-  }
-
   public ResolvedBucket resolveBucketLink(Pair<String, String> requested,
       OMClientRequest omClientRequest)
       throws IOException {
     Pair<String, String> resolved;
     if (isAclEnabled) {
-      resolved = resolveBucketLink(requested, new HashSet<>(),
+      resolved = omMReader.resolveBucketLink(requested, new HashSet<>(),
               omClientRequest.createUGI(), omClientRequest.getRemoteAddress(),
               omClientRequest.getHostName());
     } else {
-      resolved = resolveBucketLink(requested, new HashSet<>(),
+      resolved = omMReader.resolveBucketLink(requested, new HashSet<>(),
           null, null, null);
     }
     return new ResolvedBucket(requested, resolved);
-  }
-
-  public ResolvedBucket resolveBucketLink(Pair<String, String> requested)
-      throws IOException {
-
-    Pair<String, String> resolved;
-    if (isAclEnabled) {
-      UserGroupInformation ugi = getRemoteUser();
-      if (getS3Auth() != null) {
-        ugi = UserGroupInformation.createRemoteUser(
-            OzoneAclUtils.accessIdToUserPrincipal(getS3Auth().getAccessId()));
-      }
-      InetAddress remoteIp = Server.getRemoteIp();
-      resolved = resolveBucketLink(requested, new HashSet<>(),
-          ugi,
-          remoteIp,
-          remoteIp != null ? remoteIp.getHostName() :
-              omRpcAddress.getHostName());
-    } else {
-      resolved = resolveBucketLink(requested, new HashSet<>(),
-          null, null, null);
-    }
-    return new ResolvedBucket(requested, resolved);
-  }
-
-  /**
-   * Resolves bucket symlinks. Read permission is required for following links.
-   *
-   * @param volumeAndBucket the bucket to be resolved (if it is a link)
-   * @param visited collects link buckets visited during the resolution to
-   *   avoid infinite loops
-   * @param {@link UserGroupInformation}
-   * @param remoteAddress
-   * @param hostName
-   * @return bucket location possibly updated with its actual volume and bucket
-   *   after following bucket links
-   * @throws IOException (most likely OMException) if ACL check fails, bucket is
-   *   not found, loop is detected in the links, etc.
-   */
-  private Pair<String, String> resolveBucketLink(
-      Pair<String, String> volumeAndBucket,
-      Set<Pair<String, String>> visited,
-      UserGroupInformation userGroupInformation,
-      InetAddress remoteAddress,
-      String hostName) throws IOException {
-
-    String volumeName = volumeAndBucket.getLeft();
-    String bucketName = volumeAndBucket.getRight();
-    OmBucketInfo info = bucketManager.getBucketInfo(volumeName, bucketName);
-    if (!info.isLink()) {
-      return volumeAndBucket;
-    }
-
-    if (!visited.add(volumeAndBucket)) {
-      throw new OMException("Detected loop in bucket links",
-          DETECTED_LOOP_IN_BUCKET_LINKS);
-    }
-
-    if (isAclEnabled) {
-      final ACLType type = ACLType.READ;
-      checkAcls(ResourceType.BUCKET, StoreType.OZONE, type,
-          volumeName, bucketName, null, userGroupInformation,
-          remoteAddress, hostName, true,
-          getVolumeOwner(volumeName, type, ResourceType.BUCKET));
-    }
-
-    return resolveBucketLink(
-        Pair.of(info.getSourceVolume(), info.getSourceBucket()),
-        visited, userGroupInformation, remoteAddress, hostName);
   }
 
   @VisibleForTesting
